Move get_scene_module logic to extract_scene.py
